+++ packages/typespec-rust/.scripts/tspcompile.js
@@ -39,7 +39,7 @@ const cadlRanch = {
   //'cadl_duration': ['encode/duration'],
   //'cadl_bodyoptional': ['parameters/body-optionality'],
   'cadl_collectionfmt': ['parameters/collection-format'],
-  //'cadl_spread': ['parameters/spread'],
+  'cadl_spread': ['parameters/spread'],
   'cadl_contentneg': ['payload/content-negotiation'],
   'cadl_jmergepatch': ['payload/json-merge-patch'],
   //'cadl_mediatype': ['payload/media-type'],