+++ packages/typespec-rust/CHANGELOG.md
@@ -17,6 +17,7 @@
 
 * Models now derive `typespec_client_core::Model`.
 * Added support for binary responses.
+* Added support for TypeSpec spread parameters.
 
 ### Other Changes